feat #65 엔티티 자료형 수정

+++ tracky-consumer/src/main/java/kernel360/trackyconsumer/application/dto/CycleGpsRequest.java
@@ -14,13 +14,13 @@ import lombok.Setter;
 public class CycleGpsRequest {
 	private int sec;    // 발생시간 '초'
 	private String gcd;    // GPS 상태
-	private long lat;    // GPS 위도
-	private long lon;    // GPS 경도
+	private int lat;    // GPS 위도
+	private int lon;    // GPS 경도
 	private int ang;    // 방향
 	private int spd;    // 속도
-	private int sum;    // 누적주행 거리
+	private double sum;    // 누적주행 거리
 
-	public GpsHistoryEntity toGpsHistoryEntity(long maxSeq, DriveEntity drive, LocalDateTime oTime, int sum) {
+	public GpsHistoryEntity toGpsHistoryEntity(long maxSeq, DriveEntity drive, LocalDateTime oTime, double sum) {
 		return new GpsHistoryEntity(maxSeq, drive, oTime, this.gcd, this.lat, this.lon, this.ang, this.spd, sum);
 	}
 }
